# 02-YoBot-WeChat-RPA/docs/rpa-openapi-2026-09-20.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YokoWebot Control API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/license/info": {
      "get": {
        "summary": "Info",
        "operationId": "info_api_license_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/license/machine-code": {
      "get": {
        "summary": "Machine Code",
        "operationId": "machine_code_api_license_machine_code_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/license/verify": {
      "get": {
        "summary": "Verify",
        "operationId": "verify_api_license_verify_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/license/activate": {
      "post": {
        "summary": "Activate",
        "operationId": "activate_api_license_activate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/license/unbind": {
      "post": {
        "summary": "Unbind",
        "operationId": "unbind_api_license_unbind_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnbindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/api/agent/runtime-identity": {
      "get": {
        "summary": "Get Agent Runtime Identity",
        "operationId": "get_agent_runtime_identity_api_agent_runtime_identity_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Agent Runtime Identity Api Agent Runtime Identity Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/agent/contract": {
      "get": {
        "summary": "Get Agent Contract",
        "operationId": "get_agent_contract_api_agent_contract_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Agent Contract Api Agent Contract Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/runtime/capabilities": {
      "get": {
        "summary": "Get Runtime Capabilities",
        "operationId": "get_runtime_capabilities_api_runtime_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Runtime Capabilities Api Runtime Capabilities Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivationRequest": {
        "properties": {
          "activation_code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Activation Code"
          },
          "machine_code": {
            "type": "string",
            "pattern": "^[0-9A-F]{4}(?:-[0-9A-F]{4}){3}$",
            "title": "Machine Code"
          }
        },
        "type": "object",
        "required": [
          "activation_code",
          "machine_code"
        ],
        "title": "ActivationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UnbindRequest": {
        "properties": {
          "activation_code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Activation Code"
          }
        },
        "type": "object",
        "required": [
          "activation_code"
        ],
        "title": "UnbindRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}